+++ src/Database/PostgreSQL/Entity/Internal/BlogPost.hs
@@ -33,7 +33,7 @@ import GHC.Records (HasField (..))
 import Data.ByteString.Builder (byteString, char8)
 import qualified Data.List as List
 import qualified Data.Vector as Vector
-import Database.PostgreSQL.Entity (insert, insertMany)
+import Database.PostgreSQL.Entity (Field, insert, insertMany, upsert)
 import Database.PostgreSQL.Entity.Internal.QQ (field)
 import Database.PostgreSQL.Entity.Types (Entity (..), GenericEntity, PrimaryKey, TableName)
 
@@ -116,6 +116,9 @@ instance Entity BlogPost where
 -- @insertBlogPost = insert \@BlogPost@
 insertBlogPost :: BlogPost -> DBT IO ()
 insertBlogPost = insert @BlogPost
+
+upsertBlogPost :: BlogPost -> Vector Field -> DBT IO ()
+upsertBlogPost = upsert @BlogPost
 
 -- | A function to insert many blogposts at once.
 bulkInsertBlogPosts :: [BlogPost] -> DBT IO ()